+++ include/download/download_manager.h
@@ -5,7 +5,7 @@
 #include <iostream>
 #include <sstream>
 #include <string>
-#include <utils.h>
+#include <utils/utils.h>
 #include <vector>
 
 class DownloadManager {
