+++ src/download/apis/ApiTwitter.cpp
@@ -1,4 +1,4 @@
-#include <ApiTwitter.h>
+#include <download/apis/ApiTwitter.h>
 
 ApiTwitter::ApiTwitter() {
   this->_required_settings.insert(std::make_pair("api_key", ""));
