+++ src/download/download_manager.cpp
@@ -1,4 +1,4 @@
-#include <download_manager.h>
+#include <download/download_manager.h>
 
 DownloadManager::DownloadManager() {
   curl_global_init(CURL_GLOBAL_DEFAULT);
