+++ src/utils/utils.cpp
@@ -1,6 +1,6 @@
 #include <algorithm>
 #include <random>
-#include <utils.h>
+#include <utils/utils.h>
 
 std::string random_str(const unsigned int len) {
   std::random_device rd;
